Move rule graphql-naming to own file
